Allow both static and dynamic linking with raylib
By default, make will link with raylib statically. To link dynamically,
certain environment variables must be set: RAYLIB_LIBTYPE must be set to
DYNAMIC, RAYLIB_INCLUDE_DIR must be set to the directory containing the
raylib headers, and LIBS must be set to the cflags necessary to link
with raylib (such as -L<directory which contains libraylib.a> and
-lraylib, or simply the output of pkg-config --libs raylib if it's
installed.